Adding redux thunk and using it to fetch the robots.
Adding the thunk middleware and a rootReducer, containing all the reducers in the project, to the index.
Adding new constants for the different phases and possibilities when the robots are fetched.
Adding an action for fetching the robots with a catch.
Adding the reducer for managing the results of fetching the robots.
Wiring up the App by removing the original state.robots and replacing it with the action and the state.